refactor(py/executor): split builtins into scope/dirs/vars/man/script/condition/text package
